+++ src/hint-system/hint-system.ts
@@ -13,7 +13,7 @@ import { HClimbableTerrain } from './hints/climbable-terrain'
 import { HDestructible } from './hints/destructible'
 import { HEnemy, HEnemyCounter } from './hints/enemy'
 import { HProp } from './hints/prop'
-import { HOLPlatform } from './hints/rhombus-puzzle'
+import { HDynamicPlatform, HOLPlatform } from './hints/rhombus-puzzle'
 import { HMultiHitSwitch, HOneTimeSwitch, HSwitch } from './hints/switches'
 import { HDoor, HElevator, HTeleportField, HTeleportGround } from './hints/tprs'
 import { HWalls } from './hints/walls'
@@ -71,6 +71,7 @@ export class HintSystem {
         HClimbableTerrain,
         HProp,
         HAnalyzable,
+        HDynamicPlatform,
     ]
     filterType: (typeof HintTypes)[number] | 'Hints' = 'All'
     filterHintType: (typeof HintSubTypes)[number] | undefined